#144: Move extraction of filter fields to specific backend
- This makes the WHERE filter interact well with included fields for the Python ICAT backend - e.g. {'investigationUsers.id': "= '6'"}
